Refactoring code, part 1
