+++ Nidhogg/Nidhogg.vcxproj
@@ -204,7 +204,7 @@
     <ClInclude Include="FastMutex.h" />
     <ClInclude Include="FileUtils.hpp" />
     <ClInclude Include="MemoryUtils.hpp" />
-    <ClInclude Include="NidhoggUtils.h" />
+    <ClInclude Include="NidhoggCommon.h" />
     <ClInclude Include="NidhoggDeviceControl.hpp" />
     <ClInclude Include="Nidhogg.h" />
     <ClInclude Include="pch.h" />
@@ -213,12 +213,17 @@
     <ClInclude Include="WindowsTypes.hpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="AntiAnalysis.cpp" />
     <ClCompile Include="FastMutex.cpp" />
+    <ClCompile Include="FileUtils.cpp" />
+    <ClCompile Include="MemoryUtils.cpp" />
     <ClCompile Include="Nidhogg.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="ProcessUtils.cpp" />
+    <ClCompile Include="RegistryUtils.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
